Update UserRepository.java
bigint

+++ src/main/java/com/alper/springbootsmaros/repository/UserRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
+import java.math.BigInteger;
 
-public interface UserRepository extends JpaRepository<User, BigInt> {
+public interface UserRepository extends JpaRepository<User, BigInteger> {
 
     List<User> findAll();
